requirements.txt
streamlit==1.29.0
tensorflow==2.12.0
numpy
pillow
